feat: streamline admin ui
